Split out shared examples to multiple files
